moved slowDownEnemies from GameLoop to Game
