+++ src/com/codecool/snake/Globals.java
@@ -51,7 +51,4 @@ public class Globals {
         // singleton needs the class to have private constructor
     }
 
-    public GameLoop getGameLoop() {
-        return gameLoop;
-    }
 }
